+++ flake.nix
@@ -24,25 +24,25 @@
   };
 
   inputs = {
-    nix-eda.url = github:efabless/nix-eda/2.1.2;
-    libparse.url = github:efabless/libparse-python;
-    ioplace-parser.url = github:efabless/ioplace_parser;
-    volare.url = github:efabless/volare;
-    devshell.url = github:numtide/devshell;
+    nix-eda.url = "github:fossi-foundation/nix-eda/2.1.2";
+    libparse.url = "github:efabless/libparse-python";
+    ioplace-parser.url = "github:efabless/ioplace_parser";
+    ciel.url = "github:fossi-foundation/ciel";
+    devshell.url = "github:numtide/devshell";
     flake-compat.url = "https://flakehub.com/f/edolstra/flake-compat/1.tar.gz";
   };
 
-  inputs.libparse.inputs.nixpkgs.follows = "nix-eda/nixpkgs";
+  inputs.ciel.inputs.nix-eda.follows = "nix-eda";
   inputs.ioplace-parser.inputs.nix-eda.follows = "nix-eda";
-  inputs.volare.inputs.nixpkgs.follows = "nix-eda/nixpkgs";
   inputs.devshell.inputs.nixpkgs.follows = "nix-eda/nixpkgs";
+  inputs.libparse.inputs.nixpkgs.follows = "nix-eda/nixpkgs";
 
   outputs = {
     self,
     nix-eda,
     libparse,
     ioplace-parser,
-    volare,
+    ciel,
     devshell,
     ...
   }: let
@@ -53,7 +53,7 @@
     overlays = {
       default = lib.composeManyExtensions [
         (import ./nix/overlay.nix)
-        (nix-eda.flakesToOverlay [libparse ioplace-parser volare])
+        (nix-eda.flakesToOverlay [libparse ioplace-parser ciel])
         (pkgs': pkgs: {
           yosys-sby = (pkgs.yosys-sby.override { sha256 = "sha256-Il2pXw2doaoZrVme2p0dSUUa8dCQtJJrmYitn1MkTD4="; });
           klayout = (pkgs.klayout.overrideAttrs(old: {